requirements.txt
seaborn >= 0.9.0
opencv-python >= 4.4.0
numpy >= 1.18.5
scikit-learn >= 0.21.2
pyyaml >= 5.1.1